Add noble support
Adds noble support for building HTTPD.

Since there is no "full" noble builder, I have adjusted the build script to pull in `libexpat.so` from the build environment and bundle it with the HTTPD binary. This will enable the binaries to work on the base image (at least according to previous notes that say libexpat is what was missing). Bonus, this will allow us to support Jammy base as well.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>